Give a new structure to the github package
